Remove constructors for all structs
As Chromium has restrictions on the size of inlined constructors, remove
the constructors. This lines up with how pepper structs are created and used.
This means that any initialization must be done by the caller (typically
using {} to get default values or {value1, value2, ...}).

The CL also checks the size of each struct in order to ensure that the
compiler uses the same alignment as expected by the CDM, and verifies that
the struct is standard layout. It also specifiesthat each enum is an uint32_t.

BUG=none

Change-Id: If017781f5b59ecf5e554e07234bad75b0466309d