Add encryption cipher mode and pattern
In order to support alternate encryption schemes, pass the required
parameters with AudioDecoderConfig, VideoDecoderConfig, and InputBuffer
so that the CDM can determine if the encryption scheme is supported or not.

Note that the fields are added to the end of existing structures, and are
only expected to be used with the CDM 10 interface.

BUG=657957

Change-Id: If1e9ec8cf26f3ccd7117c83082cc9f0908ca934c