Add EncryptionScheme and Pattern
In order to support alternate encryption schemes, pass the required
parameters with AudioDecoderConfig, VideoDecoderConfig, and InputBuffer
so that the CDM can determine if the encryption scheme is supported or not.

The new fields are only used by the CDM 10 interface. Previous CDM
interfaces will continue to assume that any data provided is encrypted
with full sample encryption using AES-CTR mode.

BUG=657957

Change-Id: If1e9ec8cf26f3ccd7117c83082cc9f0908ca934c