Add EncryptionScheme and Pattern
In order to support alternate encryption schemes, pass the required
parameters with AudioDecoderConfig_2, VideoDecoderConfig_2, and InputBuffer_2
so that the CDM can determine if the encryption scheme is supported or not.

The new fields are only used by the CDM 10 interface. Previous CDM
interfaces will continue to assume that any data provided is encrypted
using the kCenc encryption scheme.

BUG=657957

Change-Id: If1e9ec8cf26f3ccd7117c83082cc9f0908ca934c