Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hans: Thanks for the fix and it looks good! I don't quite understand the problem though. How is cdm::FileIOClient::kSuccess in conflict with cdm::kSuccess?

cdm::FileIOClient::kSuccess is shadowing kSuccess, i.e. using the same name and thus making the outer one inaccessible.

It's not in conflict in the sense that's it's not valid, but it's generally thought a bad idea as it might lead to bugs, which is why Clang warns about it.

The warning has existed and been treated as an error in Chromium for a long time, but until recently it did not take enumerations into account. Someone recently "fixed the glitch", and now we're fixing the few instances in Chromium where the warning fires.

If it looks good to you, can I get an r+?

Patch-set: 1
Reviewer: Gerrit User 1002927 <1002927@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>